requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pdfminer.six
groq
google-api-python-client
python-multipart
chainlit
requests